requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
PyQt5>=5.15.0
torch>=1.12.0
torchvision>=0.13.0
dlib>=19.24.0
Pillow>=8.3.0
scikit-learn>=1.0.0
requests>=2.25.0
